requirements.txt
sqlglot ~= 6.2
click ~= 8.1
snowflake-connector-python